+++ src/main/java/kr/inuappcenterportal/inuportal/domain/timeTable/dto/response/timeTableItem/TimeTableMeetingResponseDto.java
@@ -2,7 +2,7 @@ package kr.inuappcenterportal.inuportal.domain.timeTable.dto.response.timeTableI
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.inuappcenterportal.inuportal.domain.course.enums.course.DayOfWeek;
+import kr.inuappcenterportal.inuportal.domain.course.enums.courseOffering.DayOfWeek;
 import kr.inuappcenterportal.inuportal.domain.course.model.CourseMeeting;
 import kr.inuappcenterportal.inuportal.domain.customSchedule.model.CustomScheduleMeeting;
