+++ src/main/java/kr/inuappcenterportal/inuportal/domain/timeTable/service/TimeTableItemService.java
@@ -1,6 +1,6 @@
 package kr.inuappcenterportal.inuportal.domain.timeTable.service;
 
-import kr.inuappcenterportal.inuportal.domain.course.enums.course.DayOfWeek;
+import kr.inuappcenterportal.inuportal.domain.course.enums.courseOffering.DayOfWeek;
 import kr.inuappcenterportal.inuportal.domain.course.model.CourseMeeting;
 import kr.inuappcenterportal.inuportal.domain.course.model.CourseOffering;
 import kr.inuappcenterportal.inuportal.domain.course.repository.CourseMeetingRepository;
